王俊:文件接口完成，未测试

+++ src/main/java/cn/edu/scnu/it/aipet/dao/UserMapper.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UserMapper {
-    public int insertUser(User user);
+    public Long insertUser(User user);
     public User getUser(@Param("iduser") Long iduser);
     public int updateUser(User role);
     public int deleteUser(@Param("iduser") Long iduser);
