+++ src/main/java/cn/edu/scnu/it/aipet/service/PetService.java
@@ -5,5 +5,5 @@ import cn.edu.scnu.it.aipet.pojo.User;
 import org.springframework.stereotype.Service;
 
 public interface PetService {
-    public int insertPet(Pet pet);
+    public Long insertPet(Pet pet);
 }
